requirements.txt
tushare==1.2.89
pyyaml==6.0
pandas>=1.3.0
numpy>=1.21.0
requests>=2.26.0